+++ src/backend/PdfBackend.cc
@@ -6,7 +6,7 @@
 
 using namespace v8;
 
-PdfBackend::PdfBackend(int width, int height)
+PdfBackend::PdfBackend(double width, double height)
   : Backend("pdf", width, height) {
   createSurface();
 }
@@ -17,7 +17,7 @@ PdfBackend::~PdfBackend() {
   destroySurface();
 }
 
-Backend *PdfBackend::construct(int width, int height){
+Backend *PdfBackend::construct(double width, double height){
   return new PdfBackend(width, height);
 }
